+++ src/initialization/init_data.c
@@ -14,16 +14,15 @@ void	ft_initialize_img_data(t_img_data *image)
 
 void  ft_init_ray_fpoint(t_ray_fpoint *ray_fpoint)
 {
-    ray_fpoint->camera_x = 0.0;
-    ray_fpoint->ray_dir_x = 0.0;
-    ray_fpoint->ray_dir_y = 0.0;
-    ray_fpoint->side_dist_x = 0.0;
-    ray_fpoint->side_dist_y = 0.0;
-    ray_fpoint->delta_dist_x = 0.0;
-    ray_fpoint->delta_dist_y = 0.0;
-    ray_fpoint->wall_dist = 0.0;
-    ray_fpoint->wall_x = 0.0;
-    
+	ray_fpoint->camera_x = 0.0;
+	ray_fpoint->ray_dir_x = 0.0;
+	ray_fpoint->ray_dir_y = 0.0;
+	ray_fpoint->side_dist_x = 0.0;
+	ray_fpoint->side_dist_y = 0.0;
+	ray_fpoint->delta_dist_x = 0.0;
+	ray_fpoint->delta_dist_y = 0.0;
+	ray_fpoint->wall_dist = 0.0;
+	ray_fpoint->wall_x = 0.0;
 }
 
 void	ft_initialize_ray_data(t_ray *ray)
@@ -45,7 +44,7 @@ void	ft_initialize_ray_data(t_ray *ray)
 	ray->draw_start = 0;
 	ray->draw_end = 0;
 	// ray->wall_x = 0;
-    ft_init_ray_fpoint(&ray->ray_distance);
+	ft_init_ray_fpoint(&ray->ray_distance);
 }
 
 void	ft_initialize_map_data(t_map_data *map_data)
@@ -70,6 +69,6 @@ void	ft_initialize_data(t_game_data *game_data)
 	game_data->texture_pixels = NULL;
 	game_data->textures = NULL;
 	ft_initialize_player_data(&game_data->player);
-    ft_initialize_map_data(&game_data->map_data);
-    ft_initialize_textures_data(&game_data->texture_data);
+	ft_initialize_map_data(&game_data->map_data);
+	ft_initialize_textures_data(&game_data->texture_data);
 }
