+++ src/player/collision_detection.c
@@ -123,42 +123,3 @@ int ft_validate_movement(t_game_data *game_data, double next_x, double next_y) {
 // 	ft_init_player_east_west(&game_data->player);
 // }
 
-//intitialize player movement
-void	ft_initialize_movemement_direction(t_player *player)
-{
-	switch (player->movement.direction)
-	{
-		case 'N':
-			player->dir_x = 0;
-			player->dir_y = -1;
-			player->plane_x = 0.66;
-			player->plane_y = 0;
-			break;
-		case 'S':
-			player->dir_x = 0;
-			player->dir_y = 1;
-			player->plane_x = -0.66;
-			player->plane_y = 0;
-			break;
-		case 'E':
-			player->dir_x = 1;
-			player->dir_y = 0;
-			player->plane_x = 0;
-			player->plane_y = 0.66;
-			break;
-		case 'W':
-			player->dir_x = -1;
-			player->dir_y = 0;
-			player->plane_x = 0;
-			player->plane_y = -0.66;
-			break;
-		default:
-			break;
-	}
-}
-
-
-void	ft_initialize_player_direction(t_game_data *game_data)
-{
-	ft_initialize_movemement_direction(&game_data->player);
-}
